Break out scale parsing before further refactoring
This probably won't fix much of the Code Climate warnings on its own,
but the changes I'll need to fix them are big enough I might as well do
this first so I'm a step closer to a proper parser
